Remove `keyring` crate in favour of file base key store
The `keyring` crate and related traits have been removed from the
kulfi_utils crate. We now use a much simpler filesystem store for iroh
secret keys.

A `malai keygen` subcommand is added to generate multiple secret keys.
Following commands take the generated key as optional arg:
- `malai http`
- `malai tcp`
- `malai folder`
If none is passed, then it tries to read the default. If the default key
does not exist then the user will have to run `malai keygen`.

NOTE: This is a breaking change as it changes the previously know
behaviour of the cli.