fix: subcmds exit with err status
All the functions to support subcommands defined by the Cli that used to
return an `eyre::Result` now handle the error themselved. The error is
handled by printing to stderr and exiting with status code 1.

This is done so that the process terminates immediately letting the user
know about the error. The graceful setup in the main function would stop
the process from existing if we simply return from them.

This goes against the process manager that we have planned to take. In a
process manager model, if a spawned task takes down the entire process,
then this is wrong. The code change done in this commit won't work with
that model. Since we are not a process manager yet and becoming a
process manager will take us some time. This grunt work of handling each
error in place is the quickest way to not let the program stuck before
it does anything meaningful.